コリジョン、描画を変更(mapでidx管理 スプライトｱﾆﾒｰｼｮﾝを変更竜 ギズモほぼ完了 fps表示変更 OBB対応

+++ Person_Base_ShibataYuki/Source/commandSpell.h
@@ -105,7 +105,7 @@ namespace Spell
 		{
 			STSpell s = st;
 			if (id == 0)
-				id = m_SpellMap.size();
+				id = static_cast<int>(m_SpellMap.size());
 			m_SpellMap.insert(std::make_pair(id, s));
 		}
